Install different zc.buildout versions.
It looks like we need different `zc.buildout` versions for running
`bootstrap.py` and `buildout`. Not doing so will create a successful
built on first retry, but will always fail on first run.

My impression is, that this is connected to version mismatch with
`setuptools`, but I am not sure though.

The current config, however, succeeded w/o any modifications for
vagrant runs on virtual Ubuntu 14.04 and 16.04.